20161230 svn

+++ SAFETYTEST/SelfTestLib/stm32f10x_STLcrc16Run.c
@@ -20,7 +20,6 @@
 *******************************************************************************/
 
 /* Includes ------------------------------------------------------------------*/
-//#include "stm32f10x_lib.h"
 #include "stm32f10x_STLlib.h"
 #include "stm32f10x_STLclassBvar.h"
 
@@ -42,9 +41,9 @@
 void STL_FlashCrc16Init(void)
 {
   p_RunCrc16Chk = ROM_START;
-  p_RunCrc16ChkInv = ((u8 *)~((u32)ROM_START));
+  p_RunCrc16ChkInv = ((u8 *)(u32)(~((u32)ROM_START)));
 
-  CurrentCrc16 = 0;
+  CurrentCrc16 = 0u;
   CurrentCrc16Inv = 0xFFFFu;
 }
 
@@ -74,13 +73,13 @@ ClassBTestStatus STL_crc16Run(void)
         CurrentCrc16 = STL_crc16(CurrentCrc16, p_RunCrc16Chk, FLASH_BLOCK);
         CurrentCrc16Inv = ~CurrentCrc16;  /* Store crc for next run */
         p_RunCrc16Chk += FLASH_BLOCK;     /* Increment pointer to next block */
-        p_RunCrc16ChkInv = ((u8 *)~((u32)p_RunCrc16Chk));
+        p_RunCrc16ChkInv = ((u8 *)(u32)(~((u32)p_RunCrc16Chk)));
         Result = TEST_RUNNING;
       }
       else
       {
         /* Do not include the ref checksum in the last block */
-        CurrentCrc16 = STL_crc16(CurrentCrc16, p_RunCrc16Chk, FLASH_BLOCK-2);
+        CurrentCrc16 = STL_crc16(CurrentCrc16, p_RunCrc16Chk, FLASH_BLOCK-2u);
         if (CurrentCrc16 == REF_CRC16)
         {
           Result = TEST_OK;
